requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-whois>=0.9.0
tldextract>=5.1.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
streamlit-aggrid>=0.3.4
pillow>=10.0.0
streamlit-lottie>=0.0.5